Remove some unnecessary functionality
YAGNI. The NetBuffer iterator had a parameter 'offset' which was unused,
I don't remember why I added it and can't think of a use case.  So I
removed it, which simplifies the implementation.
Also removed the length parameter from copy_to_slice. This was
unnecessary, as it's trivial to create another shorter slice to
accomplish the same thing (although the use case for that is also
unclear)
I did leave the length parameter for the iterator. That is not currently
used, but will be needed to implement retransmit functionality.