Refactor TCP a bit
Move packet handling out of the socket impl. This is necessary because I
need access to the lock wrapping this object in order to set timers
for things like deferred acks.
I also created a function on socket send_packet that consolidated
handling of setting all the packet fields.
There was also a copy/paste bug in one of the previous tests that broke
the TCP test in main.rs: it was calling itself recursively.