Eliminate a copy on the write path
Use writev to collect all of the buffer fragments. This was just an
interesting experiment passing data between Rust and native C
routines.
It's also possible to do this on the read path, but is a bit more
complex because the size of the packet is not known ahead of time.